hngstage4/requirements.txt
# Django Core
Django==4.2.7
djangorestframework==3.14.0

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# Database
psycopg2-binary==2.9.9

# Message Queue
pika==1.3.2

# Cache
redis==5.0.1
django-redis==5.4.0

# CORS
django-cors-headers==4.3.1

# Environment Variables
python-decouple==3.8

# Validation
phonenumbers==8.13.26

# Utilities
pytz==2023.3

# Development
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
black==23.12.0
flake8==6.1.0
isort==5.13.2

# Production Server
gunicorn==21.2.0
whitenoise==6.6.0

# Monitoring
python-json-logger==2.0.7
